requirements.txt
nextcord~=2.1
python-dotenv~=0.20
